+++ packages/suite/src/actions/wallet/discoveryActions.ts
@@ -473,7 +473,7 @@ export const start =
             }
 
             // if previous discovery status was running (typically after application start or when user added a new account)
-            // trigger fetch metadata
+            // trigger fetch metadata; necessary to load account labels
             if (discovery.status === DISCOVERY.STATUS.RUNNING) {
                 dispatch(metadataActions.fetchMetadata(deviceState));
             }