+++ packages/suite/src/services/suite/metadata/DropboxProvider.ts
@@ -11,7 +11,7 @@ import { getWeakRandomId } from '@trezor/utils';
 class DropboxProvider extends AbstractMetadataProvider {
     client: Dropbox;
     auth: DropboxAuth;
-    user: users.FullAccount | undefined;
+    user?: users.FullAccount;
     isCloud = true;
 
     constructor(token?: string) {
@@ -175,7 +175,9 @@ class DropboxProvider extends AbstractMetadataProvider {
             const account = {
                 type: this.type,
                 isCloud: this.isCloud,
-                token,
+                tokens: {
+                    refreshToken: token,
+                },
                 user: result.name.given_name,
             } as const;