+++ packages/suite/src/storage/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Storage changelog
 
+# 29
+
+-   split to `token` property for storing OAuth tokens into `tokens.accessToken` and `tokens.refreshToken`
+
 # 28
 
 -   with advent of connect v9, device.state field has changed. migrated all affected data.